internal/day25/day25_test.go
package day25

import "testing"

func TestCountFittingLockKeyPairs(t *testing.T) {
	tests := []struct {
		name  string
		input string
		want  int
	}{
		{
			name: "example",
			input: `#####
.####
.####
.####
.#.#.
.#...
.....

#####
##.##
.#.##
...##
...#.
...#.
.....

.....
#....
#....
#...#
#.#.#
#.###
#####

.....
.....
#.#..
###..
###.#
###.#
#####

.....
.....
.....
#....
#.#..
#.#.#
#####`,
			want: 3,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := CountFittingLockKeyPairs(tt.input); got != tt.want {
				t.Errorf("CountFittingLockKeyPairs() = %v, want %v", got, tt.want)
			}
		})
	}
}
